Add settings button on title bar for all screens
- Created reusable TitleBar component with settings dropdown
- Settings dropdown includes dark/light mode toggle
- Added TitleBar to all user screens (user/page.tsx)
- Added TitleBar to host/admin screens (host/page.tsx)
- Added TitleBar to projector screens (projector/page.tsx)
- Removed inline settings components in favor of TitleBar
- Cleaned up unused state variables and functions
