+++ app/projector/page.tsx
@@ -8,14 +8,12 @@ import { Id } from "@/convex/_generated/dataModel";
 import { Keypad } from "@/components/Keypad";
 import { LoadingSpinner } from "@/components/LoadingSpinner";
 import { SlideshowQuestion } from "@/components/SlideshowQuestion";
-import { Settings, Sun, Moon } from "lucide-react";
+import { TitleBar } from "@/components/TitleBar";
 
 export default function ProjectorPage() {
   const [roomCode, setClassCode] = useState("");
   const [connected, setConnected] = useState(false);
   const [storedClassCode, setStoredClassCode] = useState<string | null>(null);
-  const [settingsOpen, setSettingsOpen] = useState(false);
-  const [isDark, setIsDark] = useState(false);
 
   const room = useQuery(
     api.rooms.getRoom,
@@ -51,83 +49,10 @@ export default function ProjectorPage() {
     }
   }, [roomCode]);
 
-  // Initialize dark mode
-  useEffect(() => {
-    const stored = localStorage.getItem("theme");
-    const prefersDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
-    const shouldBeDark = stored === "dark" || (!stored && prefersDark);
-    setIsDark(shouldBeDark);
-
-    if (shouldBeDark) {
-      document.documentElement.classList.add("dark");
-    } else {
-      document.documentElement.classList.remove("dark");
-    }
-  }, []);
-
-  const toggleDarkMode = () => {
-    const newDarkMode = !isDark;
-    setIsDark(newDarkMode);
-
-    if (newDarkMode) {
-      document.documentElement.classList.add("dark");
-      localStorage.setItem("theme", "dark");
-    } else {
-      document.documentElement.classList.remove("dark");
-      localStorage.setItem("theme", "light");
-    }
-    setSettingsOpen(false);
-  };
-
-  const SettingsButton = () => (
-    <div className="fixed bottom-8 left-8 z-50">
-      <button
-        onClick={() => setSettingsOpen(!settingsOpen)}
-        className="p-3 bg-white/20 backdrop-blur hover:bg-white/30 rounded-lg transition-all duration-300"
-        aria-label="Settings"
-      >
-        <Settings className="w-5 h-5 text-white" />
-      </button>
-
-      {/* Settings Dropdown */}
-      <AnimatePresence>
-        {settingsOpen && (
-          <>
-            {/* Backdrop */}
-            <div
-              onClick={() => setSettingsOpen(false)}
-              className="fixed inset-0 z-40"
-            />
-
-            {/* Menu */}
-            <motion.div
-              initial={{ opacity: 0, y: 10, scale: 0.95 }}
-              animate={{ opacity: 1, y: 0, scale: 1 }}
-              exit={{ opacity: 0, y: 10, scale: 0.95 }}
-              className="absolute bottom-16 left-0 z-50 bg-card border border-border rounded-lg shadow-xl overflow-hidden min-w-[180px]"
-            >
-              <button
-                onClick={toggleDarkMode}
-                className="w-full px-4 py-3 flex items-center gap-3 hover:bg-muted transition-colors text-left"
-              >
-                <div className="w-5 h-5 flex items-center justify-center text-foreground">
-                  {isDark ? <Sun className="w-5 h-5" /> : <Moon className="w-5 h-5" />}
-                </div>
-                <span className="text-sm font-medium text-foreground">
-                  {isDark ? "Light Mode" : "Dark Mode"}
-                </span>
-              </button>
-            </motion.div>
-          </>
-        )}
-      </AnimatePresence>
-    </div>
-  );
-
   if (!connected) {
     return (
       <main className="flex min-h-screen flex-col items-center justify-center p-8 bg-gradient-to-br from-purple-900 to-blue-900 text-white">
-        <SettingsButton />
+        <TitleBar />
 
         <motion.div
           initial={{ opacity: 0, y: 20 }}
@@ -162,7 +87,7 @@ export default function ProjectorPage() {
   if (!room) {
     return (
       <main className="flex min-h-screen flex-col items-center justify-center p-8 bg-gradient-to-br from-purple-900 to-blue-900 text-white">
-        <SettingsButton />
+        <TitleBar />
         <LoadingSpinner size="lg" color="border-primary-300 border-t-white" />
       </main>
     );
@@ -179,7 +104,7 @@ export default function ProjectorPage() {
   if (game && game.status === "in_progress" && currentRound) {
     return (
       <main className="min-h-screen p-12 bg-background dark:bg-card text-white flex items-center justify-center">
-        <SettingsButton />
+        <TitleBar />
 
         {/* Slide Counter - Top Center */}
         <div className="fixed top-4 left-1/2 -translate-x-1/2 z-50 text-center">
@@ -240,7 +165,7 @@ export default function ProjectorPage() {
   // Phase 1
   return (
     <main className="min-h-screen p-12 bg-background dark:bg-card text-white">
-      <SettingsButton />
+      <TitleBar />
 
       <div className="max-w-7xl mx-auto space-y-12">
         {/* Header */}
